+++ CHANGELOG.md
@@ -6,10 +6,15 @@ The *alited* provides the rich and comprehensive set of features, some are uniqu
 # LAST CHANGES
 
 
-Version `v1.8.13a1 (8 Apr'26)`
+Version `v1.8.13a2 (15 Apr'26)`
 
-  - BUGFIX: no theming context menu of book-mark bar (clam theme)
+  - BUGFIX: no theming of book-mark's context menu (clam theme)
+  - BUGFIX: Saved Favorites: list of files is ruined at "Change" action
+  - BUGFIX: Saved Favorites: color of list selection isn't themed at focusing in/out
+  - NEW   : unit toolbar button: "Short/Full namespaces of procs/methods"
+  - NEW   : Saved Favorites: at selecting bar tabs, Add+ action offers them
   - NEW   : Preference/General/Test: added to preview: left tool bar & menu bar
+  - CHANGE: Projects: if current project is chosen, no actions (just close Projects)
 
 
 Version `v1.8.12 (25 Mar'26)`